lib/tgapi: Fix `-Wshorten-64-to-32` warning
When using clang as the compiler, it yields the following warning:

  tgapi.c:655:55: warning: implicit conversion loses integer precision: \
  'unsigned long' to 'int' [-Wshorten-64-to-32]

     char *escape_text = curl_easy_escape(ch, call->text, strlen(call->text));
                         ~~~~~~~~~~~~~~~~                 ^~~~~~~~~~~~~~~~~~
  1 warning generated.

It happens because the third argument of curl_easy_escape() is an
integer, but the return type of strlen() is a size_t. Fix it by casting
the return value of strlen() to int.

Fixes: ed90ca0e78e268a57412f9ade0f0b574b2e4ef03 ("lib/tgapi: Use curl_easy_escape to encode special characters")

+++ lib/tgapi.c
@@ -652,7 +652,8 @@ int tgapi_call_send_message(struct tg_api_ctx *ctx,
 	 * TODO(ammarfaizi2): Use POST method and encode the data.
 	 */
 	
-	char *escape_text = curl_easy_escape(ch, call->text, strlen(call->text));
+	char *escape_text = curl_easy_escape(ch, call->text,
+					     (int)strlen(call->text));
 	if (!escape_text)
 		return -ENOMEM;
